Nuevo modulo menu y ajustes ventas

+++ src/main/java/com/sena/chef_control/entidades/Usuario.java
@@ -18,7 +18,7 @@ public class Usuario {
 
     @Id
     @Column(name = "numero_documento")
-    private int numeroDocumento;
+    private long numeroDocumento;
 
     @Column(name = "nombre_completo")
     private String nombreCompleto;
